Support for specifying implicit dimensions for views (related to DFL-833).
Added support for specifying '*' for view width or height which allows for
setting explicit dimensions for one view and distributing remaining  dimensions
to the sibling views.
Gives better defaults for side panel for example, especially in higher
resolutions where panel could end up taking half of the screen initially.

The way it was implemented is not exhaustive - the layouts are not traversed
below first level child views. I deemed this enough for now as we presently
don't have child views that have more than one child. Can easily be extended in
the future by making added method recursive, if needed.